+++ src/client/src/pages/detail/form/stratigraphy/lithology/form/lithologyModal.tsx
@@ -69,12 +69,14 @@ export const LithologyModal: FC<LithologyEditModalProps> = ({ lithology, updateL
 
   const isUnconsolidated = formMethods.watch("isUnconsolidated");
 
-  const closeDialog = async () => {
+  const cancelDialog = () => {
+    updateLithology(lithology as Lithology, false);
+  };
+
+  const applyDialog = async () => {
     const values = getValues();
     const isValid = await formMethods.trigger();
     const hasChanges = JSON.stringify(lithology) !== JSON.stringify(values);
-
-    // Close dialog if no changes have been made or the form is valid
     if (!hasChanges || isValid) {
       prepareLithologyForSubmit(values);
       updateLithology({ ...lithology, ...values } as Lithology, hasChanges || (Boolean(lithology?.isGap) && isValid));
@@ -151,8 +153,9 @@ export const LithologyModal: FC<LithologyEditModalProps> = ({ lithology, updateL
     <FormDialog
       open={lithology !== undefined}
       title={t("lithology")}
-      onClose={closeDialog}
-      isCloseDisabled={!formState.isValid && Object.keys(formState.errors).length > 0}
+      onClose={cancelDialog}
+      onApply={applyDialog}
+      isApplyDisabled={!formState.isValid && Object.keys(formState.errors).length > 0}
       headerAction={rockTypeToggle}>
       <FormProvider {...formMethods}>
         <BoreholesCard data-cy="lithology-basic-data" title={t("basicData")}>